Move finance dashboard configs to dashboard utils
