--- src/components/DashboardSection/utils/index.ts
@@ -1,1 +1,0 @@
-export * from './charts';
