+++ src/dashboard/utils/dashboards/index.ts
@@ -1,0 +1,1 @@
+export * from './finance';
